requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.12
django-filter==21.1
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
importlib-metadata==4.11.1
iniconfig==1.1.1
Markdown==3.3.6
MarkupPy==1.14
mock==4.0.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
tomli==2.0.1
typing_extensions==4.1.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
